# Allows to turn on/off antialias for shapes.