This commit adds `nvgEnable()` and `nvgDisable()` functions for changing flags after creating the `NVGcontext` object.

Allowing changing flags is useful to get some performance gain at certain situations. For example, when applying `nvgFill()` and `nvgStroke()`  to the same path with an opaque stroke color. The antialiasing for fill operation is unnecessary because the shape edge will be covered by stroke anyway. That way, I can do something like the following code to reduce some triangle vertices from processing.

    nvgDisable(context, NVG_ANTIALIAS);
    nvgFill(context);
    nvgEnable(context, NVG_ANTIALIAS);
    nvgStroke(context);

In my real world app. I need to stroke and fill thousands of polygons per frame, and this would help improve the FPS and battery consumption a bit. Here's a simple benchmark in a test case of my app running on iPhone 6s+.

|  | Before | After|
| --- | --- | --- |
| nverts | 169,628 | 137,340 |
| FPS | 38~39 | 44~45 |