@memononen 2017-07-10
Sorry for not commenting on this earlier, I have been really busy.

I like the idea, and I think it is worth implementing. Based on your usage pattern, I think the flag should be part of `NVGstate`, though. So that you can save/restore state. That way it would be more like "drawing style", rather than backend capability.

With your implementation the shader does not change, so the flag change is not comparable to changing the creating flags. The speedup comes from not creating and rendering the extra geometry.

I suggest to change checks for `ctx->params.edgeAntiAlias` to `ctx->params.edgeAntiAlias && state->edgeAntiAlias`, enable/disable should change the state instead. Maybe call the flag `NVG_ANTIALIAS_HINT` (and make it the same value as `NVG_ANTIALIAS_HINT`, since it can only turn off antialias, not turn it on if the backend was created w/o AA.

The reason for calling it hint is that it will not work for text and maybe not strokes either. Well, device creating will not disable text AA either, still :) An alternative would be to call it something like "quick render" or something, which indicates that some corners are cut.

This change should also enable to implement a version of `drawImage` which renders just two triangles.