@olliwang 2017-07-11
Thanks @memononen. I've rewritten the commit to add a new `nvgHints()` function for setting the `NVG_ANTIALIAS_HINT` flag. Please let me know if something needs to change.

I'm not sure what do you mean about the `drawImage` part. If I understand correctly, when filling an image by the combination of `nvgRect()`, `nvgFillPaint()`, and `nvgFill()` with `edgeAntiAlias` disabled. The `glnvg__convexFill()` should kick in and use `GL_TRIANGLE_FAN` to render the quad. In my test I do see the number of `paths[i].fillCount` is equal to 4.