@memononen 2017-07-11
You did the thing I asked, but I think it does not feel quote right yet. The hint thing feels a bit premature future proofing, my bad.

This is specific optimization for the specific algorithm used by NanoVG, common to all backends
How about just calling it `nvgShapeAntialias(ctx, int enabled)`? And explain that it turns on/off antialiasing for `nvgFill` and `nvgStroke`.

There's been previous issue about rendering just sprites/images quicky. This change enables that case too, just like you describe.