@memononen 2017-07-12
One more thing :) I think the shapeAA check is missing from `nvgStroke`.