@memononen 2017-07-12
Merged, thanks for the improvement :)